+++ servus/uri.cpp
@@ -151,9 +151,10 @@ void _parseQueryMap( URIData& data )
 {
     // parse query data into key-value pairs
     std::string query = data.query;
+    data.queryMap.clear();
     while( !query.empty( ))
     {
-        const size_t nextPair = query.find( ',' );
+        const size_t nextPair = query.find( '&' );
         if( nextPair == 0 )
         {
             query = query.substr( 1 );
@@ -372,6 +373,13 @@ void URI::setPath( const std::string& path )
     _impl->getData().path = path;
 }
 
+void URI::setQuery( const std::string& query )
+{
+    URIData& data = _impl->getData();
+    data.query = query;
+    detail::_parseQueryMap( data );
+}
+
 void URI::setFragment( const std::string& fragment )
 {
     _impl->getData().fragment = fragment;
@@ -406,7 +414,7 @@ void URI::addQuery( const std::string& key, const std::string& value )
         if( data.query.empty( ))
             data.query = i->first + "=" + i->second;
         else
-            data.query += std::string( "," ) + i->first + "=" + i->second;
+            data.query += std::string( "&" ) + i->first + "=" + i->second;
     }
 }